improve perf on AMD processors, reduce casts

+++ src/MagicScaler/Magic/ConvolutionTransform.cs
@@ -9,7 +9,7 @@ namespace PhotoSauce.MagicScaler.Transforms
 	{
 		int Channels { get; }
 		int MapChannels { get; }
-		void ConvolveSourceLine(byte* istart, byte* tstart, int cb, byte* mapxstart, int smapx, int smapy);
+		void ConvolveSourceLine(byte* istart, byte* tstart, nint cb, byte* mapxstart, int smapx, int smapy);
 		void WriteDestLine(byte* tstart, byte* ostart, int ox, int ow, byte* pmapy, int smapy);
 		void SharpenLine(byte* cstart, byte* ystart, byte* bstart, byte* ostart, int ox, int ow, float amt, float thresh, bool gamma);
 	}
